refactor: reorganize imports and enhance test configurations for clarity
